+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Syntax error in request with facet queries that contain local parameters
 - HEAD requests could lead to timeouts with cURL adapter
+- Fix for reserved characters in managed resources (SOLR-6853)
 
 ### Changed